requirements.txt
pandas==2.3.1
streamlit==1.46.1
plotly==6.2.0
streamlit-authenticator==0.2.2
openpyxl==3.1.5
snowflake-connector-python==3.12.2